Improve Label with LaTeX notation
